+++ mapbox_tilesets/__init__.py
@@ -1,0 +1,3 @@
+"""mapbox_tilesets package"""
+
+__version__ = "0.3.5"
