+++ mapbox_tilesets/errors.py
