+++ tests/test_cli_list.py
@@ -5,7 +5,7 @@ from unittest import mock
 
 from click.testing import CliRunner
 
-from tilesets.scripts.cli import list
+from mapbox_tilesets.scripts.cli import list
 
 
 @pytest.mark.usefixtures("token_environ")
