+++ tests/test_cli_validate_recipe.py
@@ -4,7 +4,7 @@ import pytest
 from click.testing import CliRunner
 from unittest import mock
 
-from tilesets.scripts.cli import validate_recipe
+from mapbox_tilesets.scripts.cli import validate_recipe
 
 
 class MockResponse:
