--- tilesets/__init__.py
@@ -1,3 +1,0 @@
-"""tilesets package"""
-
-__version__ = "0.3.5"
